modules/durable_function/function_code/requirements.txt
azure-functions
azure-functions-durable
azure-identity
azure-mgmt-keyvault
requests